+++ src/main/java/com/praise/push/config/JpaAuditConfig.java
@@ -1,4 +1,4 @@
-package com.praise.praisejail.config;
+package com.praise.push.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
