+++ src/test/java/com/praise/push/PraisePushApplicationTests.java
@@ -1,4 +1,4 @@
-package com.praise.praisejail;
+package com.praise.push;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
